# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Order Transaction API",
        "version": "1.0"
    },
    "host": "",
    "basePath": "/api/v1",
    "paths": {
...
        "/category": {
            "get": {
                "tags": [
                    "/category"
                ],
                "summary": "List Category",
                "description": "List Category",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Category Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 10
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                }
            },
            "post": {
                "tags": [
                    "/category"
                ],
                "summary": "Create new Category",
                "description": "Create new Category",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "parent",
                        "in": "formData",
                        "description": "Parent Category",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Category Name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/category/{id}": {
            "get": {
                "tags": [
                    "/category"
                ],
                "summary": "View Category detail",
                "description": "View Category detail",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/category"
                ],
                "summary": "Update existing Category",
                "description": "Update existing Category",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "parent",
                        "in": "formData",
                        "description": "Parent Category",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Category Name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/category"
                ],
                "summary": "Delete existing Category",
                "description": "Delete existing Category",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/category/{id}/status": {
            "get": {
                "tags": [
                    "/category"
                ],
                "summary": "Change Category Status",
                "description": "Change Category status",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/category/mapCategoryProduct": {
            "put": {
                "tags": [
                    "/category"
                ],
                "summary": "Map Category Product",
                "description": "Map Category Product",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "formData",
                        "description": "Product ID",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "Product ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "x-access-token",
            "in": "header"
        }
    },
...
}